Update scan-site.js
More improvements, making it more error tolerant and the exports now include path (where script was executed) as well as a report that scans the content of CSS and scripts to find more references. This report is not directly something that means external requests happens (we just list external references) but it will give you somethings to investigate further.